сделал move

+++ ProjectGroup_Mars_Resource_Battle/MainProject/__astcache/d@@myprojects@mdndm@mdndm@projectgroup_mars_resource_battle@mainproject/classes@cell.h
@@ -18,7 +18,7 @@ public:
 		return type_cell;
 	}
 
-	float getHeightHill()
+	int getHeightHill()
 	{
 		return height_hill;
 	}
